+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- [Docs] Added a new step in the block's configuration
+
 ## [0.3.0] - 2020-02-17
 
 ## [0.3.0-beta] - 2020-02-14