+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
-- [Docs] Added a new step in the block's configuration
+- New step in the block's configuration
 
 ## [0.3.0] - 2020-02-17